Add missing TransitionLock to PnP handler
Otherwise we have a racing condition: if thread A is interrupted in
TunSendNetBufferLists() after the TunCheckForPause() call, thread B may
set the adapter state to TUN_STATE_PAUSING and clear the queue in
TunPnPNotifyDeviceChange(). But, the queue will get refilled by resuming
thread A and we get an adapter in paused state with a dirty NBL queue.

Signed-off-by: Simon Rozman <simon@rozman.si>