Stop closing handles on adapter halt
The "DF - PNP Surprise Remove Device Test (Reliability)" WHLK test is
failing with the following bugcheck:

Bugcheck summary:
	Code: 0x000000c4 (0x000000000000003c, 0x0000000000000378, 0x0000000000000000, 0x0000000000000000)
	Name: DRIVER_VERIFIER_DETECTED_VIOLATION
	Description: A device driver attempting to corrupt the system has been caught.  This is
because the driver was specified in the registry as being suspect (by the
administrator) and the kernel has enabled substantial checking of this driver.
If the driver attempts to corrupt the system, bugchecks 0xC4, 0xC1 and 0xA will
be among the most commonly seen crashes.
	Parameter Description:
		Parameter 1: ObReferenceObjectByHandle is being called with a bad handle.
		Parameter 2: Handle value specified by the caller.
		Parameter 3: Object type specified by the caller.
	ReportId: 93a6a499-3205-4b29-8287-7dacef889b6e

This commit disables the handle closing on adapter halt to test if this
is the cause of the problem.

Note: Device.RemoveLock no longer covers particular client handles being
open not to block TunHaltEx() infinitely.

Signed-off-by: Simon Rozman <simon@rozman.si>